:truck: move appDirs & ensureConfigFile from youwol_utils in youwol
